+++ R/experiment_functions.R
@@ -251,3 +251,23 @@ describe_experiment <- function(experiment) {
     message("Data for ", this_group, ": ", as.character(group_count))
   }
 }
+
+#' @title Remove Data from Experiment File
+#'
+#' @description Removes the data object of interest from the experiment according to
+#' a data_id. Removing the data with data_id = 12 removes the 12th dataset added to
+#' the experiment.
+#'
+#' @param experiment The experiment object
+#' @param data_id The number of the data file to be removed
+#'
+#' @return The experiment file, with the rejected call data removed.
+#'
+#' @examples \dontrun{experiment <- remove_experiment_data(experiment = experiment_object, data_id = 12)}
+#'
+#' @export
+remove_experiment_data <- function(experiment, data_id) {
+  experiment$experimental_data <- experiment$experimental_data[-as.numeric(data_id)]
+  experiment <- update_experiment(experiment)
+  return(experiment)
+}